scripts/requirements.txt
bs4
lxml
python-dateutil
pyyaml
requests

